allow backups to be disabled
