+++ module.yml
@@ -86,9 +86,9 @@ Parameters:
     Type: String
     Default: ''
   BackupRetentionPeriod:
-    Description: 'The number of days to keep backups of the table.'
+    Description: 'The number of days to keep backups of the table (set to 0 to disable).'
     Type: Number
-    MinValue: 1
+    MinValue: 0
     MaxValue: 35
     Default: 30
   BackupScheduleExpression:
@@ -108,6 +108,7 @@ Conditions:
   HasStream: !Not [!Equals [!Ref StreamViewType, 'DISABLED']]
   HasBillingAndScalingModeProvisioned: !Equals [!Ref BillingAndScalingMode, PROVISIONED]
   HasTtlAttributeName: !Not [!Equals [!Ref TtlAttributeName, '']]
+  HasBackupRetentionPeriod: !Not [!Equals [!Ref BackupRetentionPeriod, 0]]
 Resources:
   Table:
     Type: 'AWS::DynamoDB::Table'
@@ -286,11 +287,13 @@ Resources:
       OKActions:
       - 'Fn::ImportValue': !Sub '${AlertingModule}-Arn'
   BackupVault: # cannot be deleted with data
+    Condition: HasBackupRetentionPeriod
     Type: 'AWS::Backup::BackupVault'
     Properties:
       BackupVaultName: !Ref 'AWS::StackName'
       Notifications: !If [HasAlertingModule, {BackupVaultEvents: [BACKUP_JOB_STARTED, BACKUP_JOB_COMPLETED, RESTORE_JOB_STARTED, RESTORE_JOB_COMPLETED, RECOVERY_POINT_MODIFIED], SNSTopicArn: {'Fn::ImportValue': !Sub '${AlertingModule}-Arn'}}, !Ref 'AWS::NoValue']
   BackupPlan:
+    Condition: HasBackupRetentionPeriod
     Type: 'AWS::Backup::BackupPlan'
     Properties:
       BackupPlan:
@@ -304,6 +307,7 @@ Resources:
           StartWindowMinutes: 60
           TargetBackupVault: !Ref BackupVault
   BackupRole:
+    Condition: HasBackupRetentionPeriod
     Type: 'AWS::IAM::Role'
     Properties:
       AssumeRolePolicyDocument:
@@ -314,7 +318,7 @@ Resources:
             Service: 'backup.amazonaws.com'
           Action: 'sts:AssumeRole'
       Policies:
-      - PolicyName: backup
+      - PolicyName: backup # https://docs.aws.amazon.com/aws-backup/latest/devguide/access-control.html#managed-policies
         PolicyDocument:
           Version: '2012-10-17'
           Statement:
@@ -329,6 +333,7 @@ Resources:
             - 'dynamodb:DeleteBackup'
             Resource: !Sub '${Table.Arn}/backup/*'
   BackupSelection:
+    Condition: HasBackupRetentionPeriod
     Type: 'AWS::Backup::BackupSelection'
     Properties:
       BackupPlanId: !Ref BackupPlan
@@ -341,7 +346,7 @@ Outputs:
   ModuleId:
     Value: 'dynamodb-table'
   ModuleVersion:
-    Value: '1.3.0'
+    Value: '1.4.0'
   StackName:
     Value: !Ref 'AWS::StackName'
   Arn:
