+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cfn-modules/dynamodb-table",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "description": "AWS DynamoDB table with auto scaling and alerting",
   "author": "Michael Wittig <michael@widdix.de>",
   "license": "Apache-2.0",
